# Album-Figurinhas-versao-Finalzip/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Finanças Empresariais Legends 2026 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me/missions/{missionId}/start:
    post:
      operationId: startMission
      tags: [users]
      summary: Start a mission
      parameters:
        - name: missionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mission started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionProgress"
  /users/me/missions/{missionId}/submit:
    post:
      operationId: submitMissionProof
      tags: [users]
      summary: Submit evidence proof for a mission
      parameters:
        - name: missionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProofSubmission"
      responses:
        "200":
          description: Proof submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionProgress"
  /users/me/missions/{missionId}/complete:
    post:
      operationId: completeMission
      tags: [users]
      summary: Claim reward for a completed mission
      parameters:
        - name: missionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Mission completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MissionCompletionResult"
  /missions:
    get:
      operationId: listMissions
      tags: [missions]
      summary: List all missions (admin)
      responses:
        "200":
          description: Mission list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mission"
    post:
      operationId: createMission
      tags: [missions]
      summary: Create a mission (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
  /missions/{id}:
    patch:
      operationId: updateMission
      tags: [missions]
      summary: Update a mission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MissionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mission"
    delete:
      operationId: deleteMission
      tags: [missions]
      summary: Delete a mission (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Collaborator:
      type: object
      required: [id, name, role, area, management, rarity, category, points, isSpecial]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        area:
          type: string
        management:
          type: string
        photoUrl:
          type: ["string", "null"]
        yearsAtVale:
          type: ["integer", "null"]
        superPower:
          type: ["string", "null"]
        curiosity:
          type: ["string", "null"]
        achievement:
          type: ["string", "null"]
        challengeQuestion:
          type: ["string", "null"]
        challengeAnswer:
          type: ["string", "null"]
        rarity:
          type: string
          enum: [comum, rara, epica, lendaria]
        category:
          type: string
        points:
          type: integer
        isSpecial:
          type: boolean
...
    Mission:
      type: object
      required: [id, title, description, goal, rewardPoints, missionType, type, status]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        goal:
          type: integer
        rewardPoints:
          type: integer
        missionType:
          type: string
        type:
          type: string
          enum: [auto, peer, evidence]
        requiresApproval:
          type: boolean
        status:
          type: string
          enum: [open, closed]
    MissionInput:
      type: object
      required: [title, description, goal, rewardPoints]
      properties:
        title:
          type: string
        description:
          type: string
        goal:
          type: integer
        rewardPoints:
          type: integer
        missionType:
          type: string
        type:
          type: string
          enum: [auto, peer, evidence]
        requiresApproval:
          type: boolean
        status:
          type: string
          enum: [open, closed]
    MissionUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        goal:
          type: integer
        rewardPoints:
          type: integer
        missionType:
          type: string
        type:
          type: string
          enum: [auto, peer, evidence]
        requiresApproval:
          type: boolean
        status:
          type: string
          enum: [open, closed]
    MissionProgress:
      type: object
      required: [missionId, title, description, goal, progress, rewardPoints, completed, started, type, missionType, submissionStatus]
      properties:
        missionId:
          type: integer
        title:
          type: string
        description:
          type: string
        goal:
          type: integer
        progress:
          type: integer
        rewardPoints:
          type: integer
        completed:
          type: boolean
        started:
          type: boolean
        type:
          type: string
        missionType:
          type: string
        submissionStatus:
          type: string
        proofText:
          type: ["string", "null"]
        reviewNote:
          type: ["string", "null"]
    MissionCompletionResult:
      type: object
      required: [mission, bonusPoints]
      properties:
        mission:
          $ref: "#/components/schemas/MissionProgress"
        rewardCard:
          $ref: "#/components/schemas/Collaborator"
        bonusPoints:
          type: integer
    ProofSubmission:
      type: object
      required: [proofText]
      properties:
        proofText:
          type: string
...