# Album-Figurinhas-versao-Finalzip/lib/api-zod/src/generated/types/collaborator.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Finanças Empresariais Legends 2026 API
 * OpenAPI spec version: 0.1.0
 */
import type { CollaboratorRarity } from './collaboratorRarity';

export interface Collaborator {
  id: number;
  name: string;
  role: string;
  area: string;
  management: string;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  yearsAtVale?: number | null;
  /** @nullable */
  superPower?: string | null;
  /** @nullable */
  curiosity?: string | null;
  /** @nullable */
  achievement?: string | null;
  /** @nullable */
  challengeQuestion?: string | null;
  /** @nullable */
  challengeAnswer?: string | null;
  rarity: CollaboratorRarity;
  category: string;
  points: number;
  isSpecial: boolean;
}