# Album-Figurinhas-versao-Finalzip/lib/api-zod/src/generated/types/missionCompletionResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Finanças Empresariais Legends 2026 API
 * OpenAPI spec version: 0.1.0
 */
import type { Collaborator } from './collaborator';
import type { MissionProgress } from './missionProgress';

export interface MissionCompletionResult {
  mission: MissionProgress;
  rewardCard?: Collaborator;
  bonusPoints: number;
}